backend/requirements.txt
### **requirements.txt**

fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
aiohttp==3.9.1
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pillow==10.1.0
opencv-python==4.8.1.78

